Clean up root.tsx, save original code
